# server/swagger.yaml
openapi: 3.0.0
info:
  title: Project Pulse API
  version: 1.0.0
  description: API for managing users and projects.
servers:
  - url: http://localhost:3044
    description: Local development server
paths:
...
  /users/{userId}:
    get:
      summary: Get a user by ID
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    put:
      summary: Update a user by ID
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      summary: Delete a user by ID
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - Admin
            - User
        defaultAvatar:
          type: string
        avatarUrl:
          type: string
        projectsCreated:
          type: array
          items:
            type: string
        projectsMember:
          type: array
          items:
            type: string
...
    UpdateUser:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
...